+++ src/Ext/Aircraft/Body.cpp
@@ -1,6 +1,6 @@
 #include "Body.h"
+#include <Ext/TechnoType/Body.h>
 #include <Ext/WeaponType/Body.h>
-#include "..\Scenario\Body.h"
 
 // TODO: Implement proper extended AircraftClass.
 
@@ -21,4 +21,32 @@ void AircraftExt::FireBurst(AircraftClass* pThis, AbstractClass* pTarget, int sh
 			pThis->Fire(pTarget, weaponIndex);
 		}
 	}
+}
+
+// Spy plane, airstrike etc.
+bool AircraftExt::PlaceReinforcementAircraft(AircraftClass* pThis, CellStruct edgeCell)
+{
+	auto const pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType());
+	auto coords = CellClass::Cell2Coord(edgeCell);
+	coords.Z = 0;
+	AbstractClass* pTarget = nullptr;
+
+	if (pTypeExt->SpawnDistanceFromTarget.isset())
+	{
+		pTarget = pThis->Target ? pThis->Target : pThis->Destination;
+
+		if (pTarget)
+			coords = GeneralUtils::CalculateCoordsFromDistance(CellClass::Cell2Coord(edgeCell), pTarget->GetCoords(), pTypeExt->SpawnDistanceFromTarget.Get());
+	}
+
+	++Unsorted::IKnowWhatImDoing;
+	bool result = pThis->Unlimbo(coords, DirType::North);
+	--Unsorted::IKnowWhatImDoing;
+
+	pThis->SetHeight(pTypeExt->SpawnHeight.Get(pThis->Type->GetFlightLevel()));
+
+	if (pTarget)
+		pThis->PrimaryFacing.SetDesired(pThis->GetTargetDirection(pTarget));
+
+	return result;
 }